Add files via upload
## Optimization Algorithm
This code implements an optimization algorithm aimed at finding the best organism (binary string) that maximizes a fitness function. The fitness objective is to find the highest sum in the population and serves as a test for the optimizing the training set for epitope prediction

Algorithm Steps:
Initialization: The algorithm starts by generating a population of random organisms, where each organism is represented by a binary string of a specified length.

Fitness Evaluation: The fitness of each organism is calculated based on a fitness function. In this implementation, the fitness function is defined as the sum of the binary values in the organism.

Selection: The selection process involves randomly choosing parent organisms from the current population. The selection is based on a stochastic process, where organisms with higher fitness have a higher chance of being selected.

Crossover: Crossover is performed between pairs of selected parent organisms to create two new child organisms. The crossover point is randomly determined, and it allows the exchange of genetic information between parents.

Mutation: To introduce diversity into the population, a mutation operation is applied to the child organisms. With a small probability, a random bit in an organism is flipped, leading to a slight modification in its genetic makeup.

Generation Update: The new child organisms are added to the next generation's population, replacing the previous generation. This step completes one generation of the algorithm.

Termination Criteria: The algorithm continues for a fixed number of generations or until the best fitness value reaches the maximum possible fitness, which is equal to the length of the organism. This ensures that the algorithm terminates when an optimal solution is found.

Result Visualization: The algorithm records the best fitness achieved in each generation and plots the fitness history over the generations using the matplotlib library. This provides insights into the convergence behavior and improvement of the population's fitness over time.